code-viewer: move plus icon to the right side
The code viewer element must have `overflow: auto;` set in its CSS
rules to make it shrink as needed to make the grade-viewer always
visible. This causes the plus signs to the left of the code viewer
to become invisible, however. To fix this, this commit shows the
plus sign on the right side of the code viewer.