Teacher revision interface (#245)
* Add value-on and value-off props to toggle component

* Add interface for teachers revision

Adds a toggle in the top-right corner of the tree view that toggles
between the student's and the teacher's tree. Toggling it while viewing
a file that has a revision, the content of the code viewer is updated.
Revision detection is still very basic, it just links files from one
tree to the other if their paths are the same. There's not a lot else we
can do on the client side.

* Don't use <a> tag to get underline on hover

* Show teacher's revision toggle only with correct permission

* Add * to filename if it has a revision

* Show popover on files with a revision

* Pointer cursor on the asterisk on revised files

* Show asterisk after removed files

* Initial diff viewer implementation

* Show entire file diff in diff viewer instead of just the patch

* Change default code color in light theme

* Show indicator of revised files in filetree

* Sort files returned by `File.list_contents` alphabetically

* Disable line feedback on 'teacher' & 'diff' mode

* Remove language from PreferenceManager in diff mode

* Disable highlighting for large files (> 5000 lines)

* Fix issues with chrome, preference button & whitespace in diff view

* Show whitespace in CodeViewer in diff mode

* Fix duplicate directories in diff mode