Add a `with_items_in_request` decorator to check required fields
This has a few advantages over just using `ensure_keys_in_dict`. The first is
that this decorator ensures you have the required arguments in your function
declaration with the correct types. So casting is not necessary anymore and the
'casts' (they are not gone, just implicit) are checked during read time (if you
had such a thing in python). The second advantage is that as the required
objects are static (they are evaluated once, during read time), they can be
inspected more easily, it is also easier too see what different endpoints
require. This brings the third point: the documentation for request objects can
now be generated automatically or at least be checked that way (make sure every
key is described). We could even start generating a json schema from our source
code.

This PR is a proposal, so lets discuss! Please note that not all code has been
refactored as we might not even want to keep this new approach.