Speedup `/assignment/{assignment_id}/submissions/` route (#332)
* Speedup `/assignment/{assignment_id}/submissions/` route

This route was very slow because a very large amount of requests were being
made. This stops sending the `comment` field by default on this route as this
field can become quite large. Furthermore it calculates the grade way more
efficiently if the grade is a rubric grade (otherwise the grade is not
calculated but simply stored in the database) by pre loading the
`selected_items` field on the `Work` objects it also converts the
`max_rubric_points` property to a cached property as a database request is done
to lookup this field.