Fix the reload behavior of snippets (#344)
Make sure snippets are always reloaded when a user tries to expand
snippets. Also don't reload snippets on verify login as this call is done every
time a user reloads a page. Furthermore we don't retry getting snippets if this
call fails, as this not really needed and can lead (and it did) to a infinite
loop of getting snippets if a user does not have the permission to use
snippets. This PR also closes #290 as I can't reproduce it on this branch.