Add done grading notification email (#346)
* Make the way dividing and assigning works more intuitive

I tried to explain how assigning and redividing works to somebody and a lot of
small things were unclear. This commit will change that by using the following
starting point: you divide an assignment by assigning submissions, but you do
not assign a submission by dividing. This makes it more logical that the
dividing dialog is not updated when assigning (as assigning does not lead to
dividing). However we should send reminder emails to all graders that are
assigned, not divided. This commit tries to explain this with
`description-popover` elements wherever possible.

* Add initial implementation the done email

This changes the UI for setting reminders, however the backend tests for the API
are not yet updated and as the API was changed in a breaking fashion the tests
will now fail.

* Add tests and documentation for the new features

* Add forgotten new files

* Change the name of the ``change_reminder`` function

The new name better reflects what is actually does

* Fix failing travis tests

Revert the renaming of `send_reminder_mails` to `send_reminder_mail` as the
previous name was better.

* Make sure no other functions use `_send_mail` out during tests

Otherwise checking if mails are send becomes way more difficult and the other
functions are already tested.

* Fix building the documentation

* Use correct tag

* Fixed all issues raised by @olmokramer during review

The things fixed are:
- Leading underscores are removed from the docstrings as they were not present
in the names of the variables anymore.
- The default done grading template was updated to make more sense.

* Create a `cancel` method in the submit button

This makes the code to show a warning in `Notifications.vue` way cleaner. This
might also be useful in the future.

* Improve the error messages

This also creates a descriptive message when no reminder type has been chosen by
the user.

* Fix timezone issues with the reminder type

As the time returned by the server is UTC and not local time this caused the
time to decrease with an hour every time it was submitted. This now always first
parses the time returned by the server using `moment`.

* Fix more typo's as pointed out by @olmokramer

* Rename 'can_update_grading' reminder to 'can_update_course_notification'

This also fixes a bug where permissions were not checked correctly. This also
adds tests for these cases.